requirements.txt
Flask==1.1.2
marshmallow==3.7.1
mysql-connector==2.2.9
